9246 spa load/import should trust the MOS config as soon as it can 9247 open pool with missing devices 9248 zpool import -F (rewind) can not go past a config change
Reviewed by: Steve Gonczi <steve.gonczi@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Matt Ahrens <matt@delphix.com>

When opening or importing a storage pool, we trust the provided config (which may be from the zpool.cache file, or discovered from devices' labels) to various degrees. We should trust it only until we read the config from the MOS, at which point we should switch entirely to the MOS config and re-evaluate everything about the pool (e.g. how many top-level VDEVS it has). We may need to keep a small amount of info from the provided config, like the device paths.

This will allow us to be more flexible about importing damaged pools, because the provided config may not be able to be entirely accurate, but we know that the MOS config is consistent with the rest of the pool state.

We should be able to open/import a pool that has a missing top-level VDEV, if we are able to find a majority of devices. Once opened/imported, we should be able to (optionally, perhaps only when a tunable is set) continue read/write operation on the pool.

When using "zpool import -F [-T]", it doesn't allow using a config from the label which is more recent than the uberblock. The config is updated whenever a device is added, so this means that you can't "rewind" the pool to before the most recent (possibly bad) device was added.

Upstream bug: DLPX-42052 DLPX-42051 DLPX-42054